Moved the dialogs folder, (i do believe that its old location might be better but that is a discussion for later)
